moved functions and modules to lib
