+++ lib/functions/analysis.ts
