+++ lib/modules/fuzzyis.d.ts
