--- pages/modules/xlxs.d.ts
@@ -1,3 +1,0 @@
-declare module "xlsx/xlsx.mjs";
-
-declare module "xlsx/dist/cpexcel.full.mjs";
